+++ test/integration/elastic_db_admin/process_data.py
@@ -2,7 +2,7 @@
 
 """Program:  process_data.py
 
-    Description:  Integration testing of _process_data in elastic_db_admin.py.
+    Description:  Integration testing of process_data in elastic_db_admin.py.
 
     Usage:
         test/integration/elastic_db_admin/process_data.py
@@ -20,10 +20,10 @@ import unittest
 
 # Local
 sys.path.append(os.getcwd())
-import elastic_db_admin
-import lib.gen_libs as gen_libs
-import elastic_lib.elastic_class as elastic_class
-import version
+import elastic_db_admin                         # pylint:disable=E0401,C0413
+import elastic_lib.elastic_class as elcs    # pylint:disable=E0401,C0413,R0402
+import lib.gen_libs as gen_libs             # pylint:disable=E0401,C0413,R0402
+import version                                  # pylint:disable=E0401,C0413
 
 __version__ = version.__version__
 
@@ -65,7 +65,7 @@ class UnitTest(unittest.TestCase):
             self.cfg, "ssl_client_ca") else None
         self.scheme = self.cfg.scheme if hasattr(
             self.cfg, "scheme") else "https"
-        self.els = elastic_class.ElasticSearchStatus(
+        self.els = elcs.ElasticSearchStatus(
             self.cfg.host, port=self.cfg.port, user=self.user, japd=self.japd,
             ca_cert=self.ca_cert, scheme=self.scheme)
         self.els.connect()
@@ -86,7 +86,7 @@ class UnitTest(unittest.TestCase):
 
         with gen_libs.no_std_out():
             self.assertEqual(
-                elastic_db_admin._process_data(
+                elastic_db_admin.process_data(
                     self.check_list2, self.els, check_call=self.check_call,
                     cutoff_cpu=self.cutoff_cpu, cutoff_mem=self.cutoff_mem,
                     cutoff_disk=self.cutoff_disk), {})
@@ -102,7 +102,7 @@ class UnitTest(unittest.TestCase):
         """
 
         self.assertEqual(
-            elastic_db_admin._process_data(
+            elastic_db_admin.process_data(
                 self.check_list, self.els, check_call=self.check_call,
                 cutoff_cpu=self.cutoff_cpu, cutoff_mem=self.cutoff_mem,
                 cutoff_disk=self.cutoff_disk), {})