+++ test/unit/elastic_db_admin/check_status.py
@@ -22,14 +22,14 @@ import mock
 
 # Local
 sys.path.append(os.getcwd())
-import elastic_db_admin
-import lib.gen_libs as gen_libs
-import version
+import elastic_db_admin                         # pylint:disable=E0401,C0413
+import lib.gen_libs as gen_libs             # pylint:disable=E0401,C0413,R0402
+import version                                  # pylint:disable=E0401,C0413
 
 __version__ = version.__version__
 
 
-class ArgParser(object):
+class ArgParser():                                      # pylint:disable=R0903
 
     """Class:  ArgParser
 
@@ -66,7 +66,7 @@ class ArgParser(object):
         return self.args_array.get(skey, def_val)
 
 
-class ElasticSearchStatus(object):
+class ElasticSearchStatus():
 
     """Class:  ElasticSearchStatus
 
@@ -211,7 +211,7 @@ class UnitTest(unittest.TestCase):
         self.args8.args_array = {
             "-C": ["memory"], "-j": True, "-m": self.mem, "-u": self.cpu,
             "-p": self.disk}
-        self.args9.args_array = dict()
+        self.args9.args_array = {}
         self.check_call = {"memory": "chk_mem"}
         cfg = collections.namedtuple(
             "Cfg", "cutoff_mem cutoff_cpu cutoff_disk")